ETHAN READ THE COMMENT
If moving left and right is inverted after you reset imu, comment out the strafe inversion in the code.
If forward and backwards are inverted, do the inversion I did for strafe but instead for the drive variable.
Good luck at the competition! Let me know if you need anything. I won't be available after 12:00 for about an hour or two. Have fun!